+++ controllers/contacts/putContact.js
@@ -1,0 +1,26 @@
+const { BadRequest, NotFound } = require('http-errors')
+const contactsOperations = require('../../models/contacts')
+const joiSchema = require('../../middleware/validation/contactValidation')
+
+const putContact = async (req, res, next) => {
+  try {
+    const { error } = joiSchema.validate(req.body)
+    if (error) {
+      throw new BadRequest(error.message)
+    }
+    const { id } = req.params
+    const result = await contactsOperations.updateContact(id, req.body)
+    if (!result) {
+      throw new NotFound(`Contact with id=${id} not found`)
+    }
+    res.json({
+      status: 'success',
+      code: 200,
+      data: { result },
+    })
+  } catch (error) {
+    next(error)
+  }
+}
+
+module.exports = putContact
