+++ middleware/validation/contactValidation.js
@@ -1,0 +1,9 @@
+const Joi = require('joi')
+
+const joiSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().required(),
+  phone: Joi.string().min(14).max(14).required(),
+})
+
+module.exports = joiSchema
