+++ routes/api/contacts.js
@@ -1,100 +1,21 @@
 const express = require('express')
-const { NotFound, BadRequest } = require('http-errors')
-const Joi = require('joi')
 
-const contactsOperations = require('../../models/contacts')
-
-const joiSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().min(14).max(14).required(),
-})
+const getContacts = require('../../controllers/contacts/getContacts')
+const getContactById = require('../../controllers/contacts/getContactById')
+const postContact = require('../../controllers/contacts/postContact')
+const deleteContact = require('../../controllers/contacts/deleteContact')
+const putContact = require('../../controllers/contacts/putContact')
 
 const router = express.Router()
 
-router.get('/', async (_, res, next) => {
-  try {
-    const contacts = await contactsOperations.listContacts()
-    res.json({
-      status: 'success',
-      code: 200,
-      data: { contacts },
-    })
-  } catch (error) {
-    next(error)
-  }
-})
+router.get('/', getContacts)
 
-router.get('/:id', async (req, res, next) => {
-  try {
-    const { id } = req.params
-    const result = await contactsOperations.getByIdContact(id)
-    if (!result) {
-      throw new NotFound(`Contact with id=${id} not found`)
-    }
-    res.json({
-      status: 'success',
-      code: 200,
-      data: { result },
-    })
-  } catch (error) {
-    next(error)
-  }
-})
+router.get('/:id', getContactById)
 
-router.post('/', async (req, res, next) => {
-  try {
-    const { error } = joiSchema.validate(req.body)
-    if (error) {
-      throw new BadRequest(error.message)
-    }
-    const result = await contactsOperations.addContact(req.body)
-    res.status(201).json({
-      status: 'success',
-      code: 201,
-      data: { result },
-    })
-  } catch (error) {
-    next(error)
-  }
-})
+router.post('/', postContact)
 
-router.delete('/:id', async (req, res, next) => {
-  try {
-    const { id } = req.params
-    const result = await contactsOperations.removeContact(id)
-    if (!result) {
-      throw new NotFound(`Contact with id=${id} not found`)
-    }
-    res.json({
-      status: 'success',
-      code: 200,
-      massage: 'Remove success',
-    })
-  } catch (error) {
-    next(error)
-  }
-})
+router.delete('/:id', deleteContact)
 
-router.put('/:id', async (req, res, next) => {
-  try {
-    const { error } = joiSchema.validate(req.body)
-    if (error) {
-      throw new BadRequest(error.message)
-    }
-    const { id } = req.params
-    const result = await contactsOperations.updateContact(id, req.body)
-    if (!result) {
-      throw new NotFound(`Contact with id=${id} not found`)
-    }
-    res.json({
-      status: 'success',
-      code: 200,
-      data: { result },
-    })
-  } catch (error) {
-    next(error)
-  }
-})
+router.put('/:id', putContact)
 
 module.exports = router
